src/moped/moped_implementation/requirements_HiVT.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.3.0=pyhd8ed1ab_0
addict=2.4.0=pypi_0
aiohttp=3.8.1=py38h0a891b7_1
aiosignal=1.2.0=pyhd8ed1ab_0
antlr4-python3-runtime=4.8=pypi_0
argoverse=1.1.0=pypi_0
async-timeout=4.0.2=pyhd8ed1ab_0
attrs=22.1.0=pyh71513ae_1
blas=1.0=mkl
blinker=1.5=pyhd8ed1ab_0
bottleneck=1.3.5=py38h7deecbd_0
brotlipy=0.7.0=py38h0a891b7_1004
c-ares=1.18.1=h7f98852_0
ca-certificates=2022.9.24=ha878542_0
cachetools=5.2.0=pyhd8ed1ab_0
certifi=2022.9.24=pyhd8ed1ab_0
cffi=1.15.1=py38h74dc2b5_0
charset-normalizer=2.1.1=pyhd8ed1ab_0
click=8.1.3=unix_pyhd8ed1ab_2
cmake=3.24.2=pypi_0
colorama=0.4.6=pyhd8ed1ab_0
colour=0.1.5=pypi_0
contourpy=1.0.6=pypi_0
cryptography=3.4.8=py38h3e25421_1
cudatoolkit=11.1.1=ha002fc5_10
cycler=0.11.0=pypi_0
descartes=1.1.0=pypi_0
fftw=3.3.10=nompi_h77c792f_102
fonttools=4.38.0=pypi_0
frozenlist=1.3.0=py38h0a891b7_1
fsspec=2022.10.0=pyhd8ed1ab_0
future=0.18.2=pyhd8ed1ab_6
google-auth=2.14.0=pyh1a96a4e_0
google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
googledrivedownloader=0.4=pyhd3deb0d_1
grpcio=1.38.1=py38hdd6454d_0
h5py=3.7.0=pypi_0
hydra-core=1.1.0=pypi_0
idna=3.4=pyhd8ed1ab_0
imageio=2.22.3=pypi_0
importlib-metadata=5.0.0=pyha770c72_1
importlib-resources=5.10.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
isodate=0.6.1=pypi_0
jinja2=3.1.2=pyhd8ed1ab_1
joblib=1.2.0=pyhd8ed1ab_0
kiwisolver=1.4.4=pypi_0
lapsolver=1.1.0=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=12.2.0=h69a702a_19
libgfortran5=12.2.0=h337968e_19
libgomp=11.2.0=h1234567_1
libprotobuf=3.18.0=h780b84a_1
libstdcxx-ng=11.2.0=h1234567_1
libuv=1.43.0=h7f98852_0
llvmlite=0.39.1=pypi_0
markdown=3.4.1=pyhd8ed1ab_0
markupsafe=2.1.1=py38h0a891b7_1
matplotlib=3.6.2=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h95df7f1_0
mkl_fft=1.3.1=py38h8666266_1
mkl_random=1.2.2=py38h1abd341_0
mmcv=1.7.0=pypi_0
motmetrics=1.1.3=pypi_0
multidict=6.0.2=py38h0a891b7_1
ncurses=6.3=h5eee18b_3
networkx=2.8.8=pyhd8ed1ab_0
ninja=1.11.0=h924138e_0
numba=0.56.3=pypi_0
numexpr=2.8.3=py38h807cd23_0
numpy=1.19.0=pypi_0
oauthlib=3.2.2=pyhd8ed1ab_0
omegaconf=2.1.0=pypi_0
opencv-python=4.6.0.66=pypi_0
openssl=1.1.1s=h7f8727e_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.4.4=py38h6a678d5_0
pillow=9.3.0=pypi_0
pip=22.2.2=py38h06a4308_0
polars=0.14.25=pypi_0
protobuf=3.18.0=py38h709712a_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.21=pyhd8ed1ab_0
pydeprecate=0.3.1=pyhd8ed1ab_0
pyjwt=2.6.0=pyhd8ed1ab_0
pyntcloud=0.3.1=pypi_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=3.0.9=pyhd8ed1ab_0
pysocks=1.7.1=pyha2e5f31_6
python=3.8.13=haa1d7c7_1
python-dateutil=2.8.2=pyhd8ed1ab_0
python-louvain=0.15=pyhd8ed1ab_1
python_abi=3.8=2_cp38
pytorch=1.8.0=py3.8_cuda11.1_cudnn8.0.5_0
pytorch-cluster=1.5.9=py38_torch_1.8.0_cu111
pytorch-geometric=1.7.2=py38_torch_1.8.0_cu111
pytorch-lightning=1.5.2=pyhd8ed1ab_0
pytorch-scatter=2.0.8=py38_torch_1.8.0_cu111
pytorch-sparse=0.6.12=py38_torch_1.8.0_cu111
pytorch-spline-conv=1.2.1=py38_torch_1.8.0_cu111
pytz=2022.6=pyhd8ed1ab_0
pyu2f=0.1.5=pyhd8ed1ab_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0=py38h0a891b7_4
rdflib=6.2.0=pypi_0
readline=8.2=h5eee18b_0
requests=2.28.1=pyhd8ed1ab_1
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rsa=4.9=pyhd8ed1ab_0
scikit-image=0.19.3=pypi_0
scikit-learn=1.1.2=py38h6a678d5_0
scipy=1.9.1=py38h14f4228_0
setuptools=65.5.0=py38h06a4308_0
shapely=1.8.5.post1=pypi_0
six=1.16.0=pyh6c4a22f_0
sklearn=0.0=pypi_0
sqlite=3.39.3=h5082296_0
tensorboard=2.10.1=pyhd8ed1ab_0
tensorboard-data-server=0.6.0=py38h2b5fc30_2
tensorboard-plugin-wit=1.8.1=pyhd8ed1ab_0
threadpoolctl=3.1.0=pyh8a188c0_0
tifffile=2021.8.8=pypi_0
tk=8.6.12=h1ccaba5_0
torchmetrics=0.10.1=pyhd8ed1ab_0
tqdm=4.64.1=pyhd8ed1ab_0
typing-extensions=4.4.0=hd8ed1ab_0
typing_extensions=4.4.0=pyha770c72_0
urllib3=1.26.11=pyhd8ed1ab_0
werkzeug=2.2.2=pyhd8ed1ab_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.6=h5eee18b_0
yaml=0.2.5=h7f98852_2
yapf=0.32.0=pypi_0
yarl=1.7.2=py38h0a891b7_2
zipp=3.10.0=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_0
